% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Title}
\usage{
dim_reduction(data, parm, dino)
}
\arguments{
\item{data}{}

\item{parm}{}

\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
dim_reduction(complete_data, t_an, dino)
}
