% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explo-module.R
\name{explo_ui}
\alias{explo_ui}
\alias{explo_server}
\title{Shiny exploration module}
\usage{
explo_ui(id)

explo_server(id)
}
\arguments{
\item{id}{Namespace id.}
}
\value{
Shiny module.
}
\description{
Shiny exploration module
}
