% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{formula_parser}
\alias{formula_parser}
\title{Title}
\usage{
formula_parser(dat, outcome, type = "fixed", aliases = NULL, exclude = NULL)
}
\arguments{
\item{dat}{}

\item{outcome}{}

\item{type}{}

\item{aliases}{}

\item{exclude}{}
}
\description{
Title
}
