% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explo-plot.R
\name{ggcompare}
\alias{ggcompare}
\title{Dimension reduction with PCA}
\usage{
ggcompare(props, var, id, component_x = "PC1", component_y = "PC2")
}
\arguments{
\item{data}{}

\item{parm}{}

\item{...}{}
}
\description{
Dimension reduction with PCA
}
