% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfit.R, R/ggpartial.R
\name{ggpartial.last_fit}
\alias{ggpartial.last_fit}
\alias{ggpartial}
\alias{ggpartial.mc_split}
\alias{ggpartial.tune_results}
\title{ggpartial}
\usage{
\method{ggpartial}{last_fit}(
  obj,
  recipe = NULL,
  pred = NULL,
  tune = NULL,
  out,
  type = "regression",
  base_map = NULL,
  preprocessor = NULL
)

ggpartial(obj, ...)

\method{ggpartial}{mc_split}(
  obj,
  recipe,
  pred = NULL,
  tune = NULL,
  out,
  type = "regression",
  preprocessor = NULL
)

\method{ggpartial}{tune_results}(
  obj,
  recipe = NULL,
  pred = NULL,
  tune = NULL,
  out,
  type = "regression",
  base_map = NULL,
  plot_type = "dynamic",
  preprocessor = NULL
)
}
\arguments{
\item{obj}{Rsample mc_split object or tune tune_results object}

\item{x}{x variable}

\item{y}{y variable}
}
\value{

}
\description{
partial conditional regression for trained model on test data
}
