% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explo-plot.R
\name{ggtaxon}
\alias{ggtaxon}
\title{Taxon ggplot}
\usage{
ggtaxon(taxon, odds, id, cutoff = 20)
}
\arguments{
\item{id}{Identifier for}

\item{cutoff}{Cuttoff value for number of goups}

\item{Taxon}{Species abundance dataframe.}

\item{var}{Species name of interest.}
}
\description{
Taxon ggplot
}
