% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-module.R
\name{predict_ui}
\alias{predict_ui}
\title{Shiny predict module}
\usage{
predict_ui(id)
}
\arguments{
\item{id}{namespace}

\item{model_id}{reactive value containing a character string for the selected
model.}
}
\description{
Shiny predict module
}
