% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce-dims.R
\name{reduce_dims}
\alias{reduce_dims}
\title{Dimension reduction with PCA}
\usage{
reduce_dims(
  props,
  parm,
  var,
  id,
  loc,
  component_x = "PC1",
  component_y = "PC2"
)
}
\arguments{
\item{parm}{}

\item{data}{}

\item{...}{}
}
\description{
Dimension reduction with PCA
}
