% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-rmsre.R
\name{rmsre}
\alias{rmsre}
\alias{rmsre.data.frame}
\alias{rmsre_vec}
\title{Root mean square relative error}
\usage{
rmsre(data, ...)

\method{rmsre}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

rmsre_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{}

\item{...}{}
}
\description{
Root mean square relative error
}
