% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{role_organizer}
\alias{role_organizer}
\title{Title}
\usage{
role_organizer(
  dat,
  outcome,
  group = c("sample_id", "hole_id"),
  temporal = c("sample_depth_mbsf", "age_ma"),
  spatial = c("longitude", "latitude"),
  meta = c("site_hole", "source_citation"),
  aliases = NULL
)
}
\arguments{
\item{dat}{}

\item{outcome}{}

\item{group}{}

\item{temporal}{}

\item{spatial}{}

\item{meta}{}

\item{aliases}{}
}
\description{
Title
}
