% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-taxo.R
\name{step_taxo}
\alias{step_taxo}
\title{Reduce taxonomic depth}
\usage{
step_taxo(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  skip = FALSE,
  id = recipes::rand_id("taxo")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}

\item{options}{A list of options to the default method for
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}. Argument defaults are set to \verb{names_to = "names}
and \code{values_to = "values"}.}
}
\value{
step
}
\description{
\code{step_taxa} creates a \emph{specification} of a recipe
step that will reduce the taxonomic depth.
}
