% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-zerogeodist.R
\name{step_zerogeodist}
\alias{step_zerogeodist}
\title{Filter Minimum Distance for Spatial Coordinates}
\usage{
step_zerogeodist(
  recipe,
  lon = NULL,
  lat = NULL,
  role = "spatial",
  trained = FALSE,
  options = list(min_dist = 2500),
  skip = FALSE,
  id = recipes::rand_id("zerogeodist")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{lon, lat}{Selector functions to choose which variables are
used by the step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{options}{A list of options to the default method for
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}. Argument defaults are set to \verb{names_to = "names}
and \code{values_to = "values"}.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
step
}
\description{
\code{step_zerogeodist} creates a \emph{specification} of a recipe
step that will pivot data for multilevel models.
}
