% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-plot.R
\name{taxon_plot}
\alias{taxon_plot}
\title{Taxon ggplot}
\usage{
taxon_plot(taxon, var, id, cutoff = 30)
}
\arguments{
\item{var}{Species name of interest.}

\item{id}{Identifier for}

\item{cutoff}{Cuttoff value for number of rgoups}

\item{Taxon}{Species abundance dataframe.}
}
\description{
Taxon ggplot
}
