% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferice-package.R
\docType{package}
\name{transferice-package}
\alias{transferice}
\alias{transferice-package}
\title{transferice: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Martin Schobben \email{schobbenmartin@gmail.com} (\href{https://orcid.org/0000-0001-8560-0037}{ORCID})

}
\keyword{internal}
