% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferice_app.R
\name{transferice_app}
\alias{transferice_app}
\title{Launch the transferice app}
\usage{
transferice_app(port = 5912)
}
\arguments{
\item{port}{Numeric indicating the port number.}
}
\value{
Interactive shiny session in localhost
}
\description{
The video animations of this app are in mkv format which is currently only
supported by the latest versions for the Google Chrome browser.
}
