% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard-module.R
\name{wizard_ui}
\alias{wizard_ui}
\alias{wizard_server}
\title{Wizard module}
\usage{
wizard_ui(id, tabs)

wizard_server(id, n)
}
\arguments{
\item{id}{Namespace of the module.}

\item{tabs}{\code{Taglist} with elements to be included within the tabs. The
number of elements determines the number of slides in the wizard.}
}
\description{
The module creates a wizard that takes you through a selection of slides.
}
