% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfit.R
\name{ggfit}
\alias{ggfit}
\title{R squared plot}
\usage{
ggfit(final, parms, averaging = "an", selected = "all")
}
\arguments{
\item{final}{}

\item{parms}{}

\item{averaging}{}

\item{selected}{}
}
\value{

}
\description{
R squared plot
}
