% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfit.R, R/ggpartial.R
\name{ggpartial.last_fit}
\alias{ggpartial.last_fit}
\alias{ggpartial}
\alias{ggpartial.mc_split}
\alias{ggpartial.tune_results}
\title{ggpartial}
\usage{
\method{ggpartial}{last_fit}(
  obj,
  workflow = NULL,
  pred = NULL,
  tune = NULL,
  out,
  type = "regression",
  base_map = NULL,
  height = NULL,
  width = NULL
)

ggpartial(obj, ...)

\method{ggpartial}{mc_split}(
  obj,
  workflow,
  pred = NULL,
  tune = NULL,
  out,
  type = "regression",
  base_map = NULL,
  return_type = "plot",
  height = NULL,
  width = NULL
)

\method{ggpartial}{tune_results}(
  obj,
  workflow = NULL,
  pred = NULL,
  tune = NULL,
  out,
  type = "regression",
  base_map = NULL,
  plot_type = "dynamic",
  mc_cores = 4,
  renderer = "mkv",
  height = NULL,
  width = NULL
)
}
\arguments{
\item{obj}{Rsample mc_split object or tune tune_results object.}

\item{pred}{Predictor variable (character string).}

\item{tune}{Tune variable (character or numeric string).}

\item{out}{Outcome variable (character or symbol).}

\item{type}{Plot output type. "Regression" for a simple scatter plot with a
regression line. "Spatial" for a projection on a global map.}

\item{recipe}{Recipe for feature engineering (recipe object).}

\item{preprocessor}{Character string detailing the engineering steps.}
}
\value{
A \code{ggplot2:\link[ggplot2:ggplot()]{ggplot}} or pathway to a plot
or animation.
}
\description{
This function plots several aspect of the machine learning workflow.
}
\details{
\itemize{
\item Feature engineering for subsets of variables of the data.
\item Partial conditional regression for trained models on the test data split.
\item R squared plot for the final fit on all variables.
}
}
