% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpartial.R
\name{ggpartial}
\alias{ggpartial}
\title{ggpartial}
\usage{
ggpartial(partials, fold, tune, pred, type = "regression", base_map = NULL)
}
\arguments{
\item{model}{Parsnip model object}

\item{x}{x variable}

\item{y}{y variable}
}
\value{

}
\description{
partial conditional regression for trained model on test data
}
