% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-print.R
\name{print_predict}
\alias{print_predict}
\title{print HTML model setup
Model print}
\usage{
print_predict(obj, new_data, out)
}
\arguments{
\item{obj}{Character string identifying the finalized workflow.}

\item{new_data}{The dataframe used for the prediction.}

\item{out}{The parameter of interest.}
}
\value{
HTML text
}
\description{
print HTML model setup
Model print
}
